Move onto case class methods for a more concise API
