+++ app/model/Check.scala
@@ -4,14 +4,6 @@ import scala.collection.JavaConverters._
 import net.logstash.logback.marker.Markers
 import play.api.libs.json.{Json, Reads}
 
-object TextRange {
-  implicit val reads = Json.reads[TextRange]
-}
-
-case class TextRange(from: Int, to: Int) {
-  def length = to - from
-}
-
 case class Check(
   documentId: Option[String],
   requestId: String,
