+++ app/model/RuleMatch.scala
@@ -49,4 +49,21 @@ case class RuleMatch(rule: BaseRule,
                      replacement: Option[Suggestion] = None,
                      markAsCorrect: Boolean = false,
                      matchContext: String,
-                     matcherType: String)
+                     matcherType: String) {
+  /**
+    * Map the range this match applies to through the given ranges, adjusting its range accordingly.
+    */
+  def mapThroughIgnoredRanges(ignoredRanges: List[TextRange]): RuleMatch = {
+    val (newMatch, _) = ignoredRanges.foldLeft((this, List.empty[TextRange]))((acc, range) => acc match {
+      case (ruleMatch, rangesAlreadyApplied) => {
+        val mappedRange = rangesAlreadyApplied.foldRight(range)((acc, range) => acc.mapAddedRange(range))
+        val newMatchRange = TextRange(ruleMatch.fromPos, ruleMatch.toPos).mapAddedRange(mappedRange)
+        val newRuleMatch = ruleMatch.copy(fromPos = newMatchRange.from, toPos = newMatchRange.to)
+        (newRuleMatch, rangesAlreadyApplied :+ mappedRange)
+      }
+    })
+
+    newMatch
+  }
+}
+
